@@ -1,21 +1,17 @@
-import React from 'react'
-import SearchInput from './components/SearchInput'
-import {BookType} from './components/Book'
-import Books from './components/Books'
+import { Routes, Route } from 'react-router-dom'
 
-interface Response {
-  data?: {
-    items: BookType[]
-  }
-}
+import Home from './containers/Home'
+import BookStore from './containers/BookStore'
+import BookDetail from './containers/BookDetail'
 
 const App = () =>  {
-  const [response, setResponse] = React.useState<Response>({})
   return (
-    <div>
-      <SearchInput setResponse={setResponse} />
-	{response.data && <Books books={response.data.items} /> }
-    </div>
+    <Routes>
+      <Route element={<Home />} index />
+      <Route element={<BookStore />} path='/bookstore' />
+      <Route element={<BookDetail />} path='/bookstore/:bookId' />
+      <Route element={<h1>404 not found</h1>} path="*" />
+    </Routes>
   )
 }
 
